+++ experiments/integrated/performance/profile_allocations.jl
@@ -229,7 +229,10 @@ soilco2_type = Soil.Biogeochemistry.SoilCO2Model{FT}
 soilco2_ps = SoilCO2ModelParameters(FT)
 
 # soil microbes args
-Csom = ClimaLand.PrescribedSoilOrganicCarbon{FT}(TimeVaryingInput((t) -> 5))
+subsurface_space = soil_domain.space.subsurface
+Csom = ClimaLand.PrescribedSoilOrganicCarbon{FT}(
+    ClimaCore.Fields.zeros(subsurface_space) .+ 5,
+)
 
 # Set the soil CO2 BC to being atmospheric CO2
 soilco2_top_bc = Soil.Biogeochemistry.AtmosCO2StateBC()