+++ experiments/long_runs/land.jl
@@ -408,7 +408,9 @@ function setup_prob(t0, tf, Δt; outdir = outdir, nelements = (101, 15))
     soilco2_type = Soil.Biogeochemistry.SoilCO2Model{FT}
 
     # soil microbes args
-    Csom = ClimaLand.PrescribedSoilOrganicCarbon{FT}(TimeVaryingInput((t) -> 5))
+    Csom = ClimaLand.PrescribedSoilOrganicCarbon{FT}(
+        ClimaCore.Fields.zeros(subsurface_space) .+ 5,
+    )
 
     # Set the soil CO2 BC to being atmospheric CO2
     soilco2_top_bc = Soil.Biogeochemistry.AtmosCO2StateBC()